// github/types/paths/orgs/{org}/repos.types.ts
export type reposListForOrg = (
  $: OmitValueWhenNever<{
    query: reposListForOrg_Query;
    querystring: never;
    path: reposListForOrg_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: {
          Link: { schema: string };
        };
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: Array<minimal_repository>;
          };
        };
        examples: {
          default: unknown;
        };
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;
// ...
export type reposListForOrg_Query = {
  /**
   * Specifies the types of repositories you want returned.
   */
  type?: "all" | "public" | "private" | "forks" | "sources" | "member";
  /**
   * The property to sort the results by.
   */
  sort?: "created" | "updated" | "pushed" | "full_name";
  /**
   * The order to sort by. Default: `asc` when using `full_name`, otherwise `desc`.
   */
  direction?: "asc" | "desc";
  per_page?: number;
  page?: number;
};

export type reposListForOrg_Path = { org: string };